menambahkan migrasi skema transaksi

+++ app/Database/Migrations/2026-06-25-101503_CreateTransactionTable.php
@@ -4,10 +4,14 @@ namespace App\Database\Migrations;
 
 use CodeIgniter\Database\Migration;
 
-class CreateTransactionTable extends Migration
+class CreateTransactionTableCurrent extends Migration
 {
     public function up()
     {
+        if ($this->db->tableExists('transaction')) {
+            return;
+        }
+
         $this->forge->addField([
             'id'          => [
                 'type'           => 'INT',
@@ -54,6 +58,8 @@ class CreateTransactionTable extends Migration
 
     public function down()
     {
-        $this->forge->dropTable('transaction');
+        if ($this->db->tableExists('transaction')) {
+            $this->forge->dropTable('transaction');
+        }
     }
 }
